bookstore-microservice/cart-service/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
requests==2.32.3
djangorestframework-simplejwt==5.3.1
